requirements.txt
discord.py[speed]
google-cloud-dialogflow
requests==2.23.0
cryptography
firebase-admin
